requirements.txt
google-generativeai>=0.3.1
chromadb
sentence-transformers
langchain
python-dotenv
gradio
pandas
numpy
openai
torch
transformers
